pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qtrader"
version = "0.0.1"
description = "Lightweight kdb+ trading platform built on qi"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
]

[project.scripts]
qtrader = "qtrader.cli.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]